+++ test/generators/active_storage_generators/aws_active_storage_install_generator_test.rb
@@ -3,7 +3,7 @@
 require "test_helper"
 require "generators/boring/active_storage/google/install/install_generator"
 
-class GoogleActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
+class AwsActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
   tests Boring::ActiveStorage::Google::InstallGenerator
   setup :build_app
   teardown :teardown_app
@@ -14,16 +14,16 @@ class GoogleActiveStorageInstallGeneratorTest < Rails::Generators::TestCase
     app_path
   end
 
-  def test_should_configure_google_for_active_storage
+  def test_should_configure_aws_for_active_storage
     Dir.chdir(app_path) do
       quietly { run_generator }
 
       assert_file "Gemfile" do |content|
-        assert_match(/google-cloud-storage/, content)
+        assert_match(/aws-sdk-s3/, content)
       end
 
       assert_file "config/environments/production.rb" do |content|
-        assert_match("config.active_storage.service = :google", content)
+        assert_match("config.active_storage.service = :amazon", content)
         assert_no_match("config.active_storage.service = :local", content)
       end
     end